Modifying nginx config to better serve up Angular application (#2)
Found this solution for serving up an Angular app where we want to make sure index.html is not cached by browsers. The solution also uses try_files, which we were not using before, which makes it nicer for users because a URL like "http://<host>/logs" will actually respond with index.html and Angular will start on the Logs page. Previously this would have resulted in a 404.

* Fixing error_page config; /404.html is not on the nginx image /50x.html is

* Better serve up Angular content; make sure index.html is not cached

* Turning on gzip compression of static content